fix(refactor): avoid TDZ cycles via utils barrel for new helper imports
The previous commits introduced fresh imports from `src/utils/index.js` in
three modules that participate in barrel-induced TDZ cycles via
`ConfigurationSchema.ts`:

- src/charging-station/ui-server/UIServerNet.ts is imported by
  ConfigurationSchema.ts, so importing back through utils/index.js produces
  `ReferenceError: Cannot access 'isHostLiteralWithoutPort' before
  initialization`.
- src/worker/WorkerAbstract.ts is exported via worker/index.js, which is
  consumed by ConfigurationSchema.ts (`WorkerProcessType`), so importing
  back through utils/index.js breaks Zod enum initialization with
  `TypeError: Cannot convert undefined or null to object` in
  `getEnumValues`.
- src/charging-station/TemplateSchema.ts uses the same cycle-prone barrel
  defensively, preempting future Zod-init breakage.

Resolution: import `convertToInt` and `isNotEmptyString` directly from
`./Utils.js` (the leaf module), mirroring the repo-established
cycle-avoidance pattern at `src/utils/ConfigurationMigrations.ts:3-4`
(`import { BaseError } from '../exception/BaseError.js'` with the same kind
of inline note).

Tests: previously failing
- tests/charging-station/ui-server/UIServerNet.test.ts
- tests/worker/WorkerUtils.test.ts
both pass; full suite 2857/2857.
