+++ src/charging-station/ui-server/UIServerNet.ts
@@ -1,6 +1,9 @@
 import { isIP } from 'node:net'
 
-import { convertToInt } from '../../utils/index.js'
+// Direct path: the `utils/index.js` barrel re-exports ConfigurationSchema.ts
+// which itself imports from this module — using the barrel triggers a TDZ
+// cycle (`Cannot access 'isHostLiteralWithoutPort' before initialization`).
+import { convertToInt } from '../../utils/Utils.js'
 
 export const LOOPBACK_HOSTNAME = 'localhost'
